product/structure: allow strings in Product::amount
This is a workaround for JavaScript: in JS there cannot be any
trailing zeros after the decimal point in a number, so POSTing a
number like 1.0 is impossible. On the other hand, Serde cannot
deserializes a positive number into an i64 in mongodb::Bson. Thus,
we can only support using strings in Product::amount. Jesus.